refactor: 폴더구조 개선
