+++ src/Root.tsx
@@ -1,6 +1,6 @@
 import GlobalStyle from "style/global";
 import { ThemeProvider } from "styled-components";
-import { darkTheme, lightTheme } from "style/theme";
+import { lightTheme } from "style/theme";
 import App from "components/App";
 import { RecoilRoot } from "recoil";
 import "react-toastify/dist/ReactToastify.css";
