+++ src/components/Router/index.tsx
@@ -1,20 +1,20 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Header from "../Header";
+import Header from "../common/Header";
 import Main from "../Main";
-import Nav from "../Nav";
-import MyMusicStore from "components/MyMusicDetail";
-import TodayMusicDetail from "components/TodayMusicDetail";
-import PendingMusicDetail from "components/PendingMusicDetail";
+import Nav from "../common/Nav";
+import MyMusic from "components/MyMusic";
+import TodayMusic from "components/TodayMusic";
+import PendingMusic from "components/PendingMusic";
 
 const Router = () => {
   return (
-    <BrowserRouter basename="/wakesong">
+    <BrowserRouter>
       <Header />
       <Routes>
         <Route path="/" element={<Main />} />
-        <Route path="/mymusicdetail" element={<MyMusicStore />} />
-        <Route path="/todaymusicdetail" element={<TodayMusicDetail />} />
-        <Route path="/pendingmusicdetail" element={<PendingMusicDetail />} />
+        <Route path="/mymusic" element={<MyMusic />} />
+        <Route path="/todaymusic" element={<TodayMusic />} />
+        <Route path="/pendingmusic" element={<PendingMusic />} />
       </Routes>
       <Nav />
     </BrowserRouter>
