+++ src/hooks/melonChart/useMelonChart.ts
@@ -13,14 +13,15 @@ const useMelonChart = () => {
       {
         onSuccess: (e) => {
           if (e.status === 226) {
-            toast.error("이미 이번주에 기상송을 신청하셨습니다");
+            toast.error("이미 이번주에 기상송을 신청했습니다!");
             return;
           }
-          toast.success("신청 성공");
+
+          toast.success("기상송을 신청했습니다!");
         },
 
         onError: () => {
-          toast.error("신청 실패");
+          toast.error("기상송 신청을 실패했습니다!");
         },
       }
     );
