--- src/repository/log/log.param.ts
@@ -1,4 +1,0 @@
-export interface postModuleLogParam {
-  description: string;
-  moduleName: string;
-}
