--- src/repository/log/log.repository.ts
@@ -1,17 +1,0 @@
-import { postModuleLogParam } from "./log.param";
-import dodamAxios from "lib/axios/dodamAxios";
-
-class LogRepoisitory {
-  public async postModuleLog({
-    description,
-    moduleName,
-  }: postModuleLogParam): Promise<void> {
-    await dodamAxios.post("/logging/function", {
-      description,
-      moduleName,
-      platform: "WEB",
-    });
-  }
-}
-
-export default new LogRepoisitory();
